💄refactor(datatypes): isolate CRDT execution and remove test operation
