+++ docs/transaction-and-rollback.md
@@ -163,4 +163,4 @@ When adding a new local `OperationBody` variant, two local responsibilities must
 1. `execute_local_operation` — locate the target once, capture the state required for rollback, apply the change, and return `LocalOperationOutcome`
 2. `apply_rollback_action` — restore the captured state during rollback
 
-An `Err` from local execution must leave the CRDT unchanged because no rollback action is recorded for a failed operation. Remote execution uses `execute_remote_operation` returning `Result<()>`; it never creates rollback actions. Local execution, remote execution, and rollback application are implemented at the concrete CRDT level (for example, `CounterCrdt`). The `Crdt` enum wrapper in `src/datatypes/crdts/mod.rs` lifts each concrete action into `RollbackAction` and owns the single mismatch check.
+An `Err` from local execution must leave the CRDT unchanged because no rollback action is recorded for a failed operation. Remote execution uses `execute_remote_operation` returning `Result<()>`; it never creates rollback actions. Local execution, remote execution, and rollback application are implemented at the concrete CRDT level (for example, `CounterCrdt`). The `Crdt` enum wrapper in `src/datatypes/crdts/crdt.rs` routes each top-level `RollbackAction` to the matching concrete CRDT and owns the single mismatch check.
