+++ src/datatypes/crdts/counter_crdt.rs
@@ -151,9 +151,7 @@ mod tests_counter_crdt {
             let context = OperationContext::new(&operation, &Cuid::default());
             let outcome = counter.execute_local_operation(&context).unwrap();
             let (return_value, action) = outcome.into_parts();
-            let RollbackAction::Counter(action) = action else {
-                panic!("counter returned a non-counter rollback action");
-            };
+            let RollbackAction::Counter(action) = action;
 
             assert!(matches!(return_value, ReturnType::Counter(value) if value == expected_value));
             assert_eq!(counter.value(), expected_value);
